# old_python_fastapi_backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeRoute API",
    "description": "SafeRoute finds safer routes through Southampton by analyzing UK Police crime data.\n\nFeatures: Safety heatmaps with crime risk visualization, multiple route alternatives ranked by safety score, historical crime analytics with time-of-day weighting, Redis caching, and automated monthly data updates.\n\nData sources: UK Police API for crime data, OpenRouteService for routing, H3 hexagonal spatial indexing for efficient geographic queries.\n\nAuthentication is optional. Authenticated users get route history tracking, custom preferences, and higher rate limits. Use /auth/login to obtain a JWT access token.\n    ",
    "contact": {
      "name": "SafeRoute API Support",
      "url": "https://github.com/ss2d22/saferoute"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me/history": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User History",
        "description": "Get user's route history with pagination and filters.",
        "operationId": "get_user_history_api_v1_users_me_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User History",
        "description": "Delete all user history.",
        "operationId": "delete_user_history_api_v1_users_me_history_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/history/{history_id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete History Item",
        "description": "Delete a single history item.",
        "operationId": "delete_history_item_api_v1_users_me_history__history_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "history_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "History Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/settings": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User Settings",
        "description": "Get user settings.",
        "operationId": "get_user_settings_api_v1_users_me_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update User Settings",
        "description": "Update user settings.",
        "operationId": "update_user_settings_api_v1_users_me_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User Account",
        "description": "Delete user account (requires password confirmation).",
        "operationId": "delete_user_account_api_v1_users_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteAccountRequest": {
        "properties": {
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "DeleteAccountRequest",
        "description": "Delete account request (requires password confirmation)."
      },
      "DeleteHistoryResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "deleted_count": {
            "type": "integer",
            "title": "Deleted Count"
          }
        },
        "type": "object",
        "required": [
          "message",
          "deleted_count"
        ],
        "title": "DeleteHistoryResponse",
        "description": "Response for history deletion."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RouteHistoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "HistoryListResponse",
        "description": "Paginated history list response."
      },
...
      "RouteHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "origin": {
            "additionalProperties": true,
            "type": "object",
            "title": "Origin"
          },
          "destination": {
            "additionalProperties": true,
            "type": "object",
            "title": "Destination"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "safety_score_best": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Score Best"
          },
          "distance_m_best": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M Best"
          },
          "duration_s_best": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S Best"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "origin",
          "destination",
          "mode",
          "safety_score_best",
          "distance_m_best",
          "duration_s_best"
        ],
        "title": "RouteHistoryItem",
        "description": "Single route history item."
      },
...
      "UserSettings": {
        "properties": {
          "history_enabled": {
            "type": "boolean",
            "title": "History Enabled",
            "default": true
          },
          "history_retention_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "History Retention Days",
            "default": 90
          },
          "default_safety_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Default Safety Weight",
            "default": 0.8
          },
          "default_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Mode"
          }
        },
        "type": "object",
        "title": "UserSettings",
        "description": "User settings."
      },
      "UserSettingsUpdate": {
        "properties": {
          "history_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Enabled"
          },
          "history_retention_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "History Retention Days"
          },
          "default_safety_weight": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Safety Weight"
          },
          "default_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Mode"
          }
        },
        "type": "object",
        "title": "UserSettingsUpdate",
        "description": "User settings update request."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}